rename ores package to ore
